+++ 1.3/Defs/EmpireMainTabDefs/EmpireMainTabDefs.xml
@@ -10,19 +10,19 @@
 
 	<!-- ~~~~~~~~~~~ Actual Tabs ~~~~~~~~~~~ -->
 
-	<Empire_Rewritten.Windows.EmpireMainTabDef>
+	<Empire_Rewritten.Windows.MainTabWindowTabs.MainTabWindowTabDef>
 		<defName>Main</defName>
-		<tabClass>Empire_Rewritten.Windows.OverviewTabs.MainTab</tabClass>
+		<tabClass>Empire_Rewritten.Windows.MainTabWindowTabs.OverviewTab</tabClass>
 		<label>Empire_TabName_Main</label>
 		<description>Overview over all the empire stuff plz change this string later</description>
 		<order>0</order>
-	</Empire_Rewritten.Windows.EmpireMainTabDef>
+	</Empire_Rewritten.Windows.MainTabWindowTabs.MainTabWindowTabDef>
 
-	<Empire_Rewritten.Windows.EmpireMainTabDef>
+	<Empire_Rewritten.Windows.MainTabWindowTabs.MainTabWindowTabDef>
 		<defName>Storage</defName>
-		<tabClass>Empire_Rewritten.Windows.OverviewTabs.StorageTab</tabClass>
+		<tabClass>Empire_Rewritten.Windows.MainTabWindowTabs.StorageTab</tabClass>
 		<label>Empire_TabName_Storage</label>
 		<description>Check your Empire's currently stored materials this string is also WIP</description>
 		<order>1</order>
-	</Empire_Rewritten.Windows.EmpireMainTabDef>
+	</Empire_Rewritten.Windows.MainTabWindowTabs.MainTabWindowTabDef>
 </Defs>
